refactor(folder): Change dir name from lower case to upper case
