--- src/@common/highlighter/index.ts
@@ -1,5 +1,0 @@
-import Highlighter from './Highlighter';
-
-export {
-  Highlighter
-};
