+++ src/home/pages/react/ExBasicSyntax/index.ts
@@ -1,0 +1,3 @@
+import { ExBasicSyntax } from './ExBasicSyntax';
+
+export { ExBasicSyntax };
