+++ src/home/pages/react/ExCart/index.ts
@@ -1,0 +1,3 @@
+import { ExCart } from './ExCart';
+
+export { ExCart };
