+++ src/home/pages/react/ExProps/index.ts
@@ -1,0 +1,3 @@
+import { ExProps } from './ExProps';
+
+export { ExProps };
