+++ src/home/pages/react/ExRedux/index.ts
@@ -1,0 +1,3 @@
+import { ExRedux } from './ExRedux';
+
+export { ExRedux };
