+++ src/home/pages/react/ExReduxCounter/index.ts
@@ -1,0 +1,3 @@
+import { ExReduxCounter } from './ExReduxCounter';
+
+export { ExReduxCounter };
