+++ src/home/pages/react/ExReduxTodo/index.ts
@@ -1,0 +1,3 @@
+import { ExReduxTodo } from './ExReduxTodo';
+
+export { ExReduxTodo };
