+++ src/home/pages/react/ExState/ExState.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 // import StateClass from './StateClass';
 import {StateFc} from './StateFc';
-import {Highlighter} from '@common/highlighter';
+import Highlighter from '@common/Highlighter';
 import {StateClassCode, StateFcCode} from '@common/code/react/ExStateCode';
 
 export const ExState = () => {
