+++ src/home/pages/react/ExTodoList/index.ts
@@ -1,0 +1,3 @@
+import { ExTodoList } from './ExTodoList';
+
+export { ExTodoList };
