+++ src/home/pages/react/ExVeneer/index.ts
@@ -1,0 +1,3 @@
+import { ExVeneer } from './ExVeneer';
+
+export { ExVeneer };
