+++ src/home/pages/react/Welcome.tsx
@@ -1,6 +1,6 @@
 import styled from 'styled-components'
 import React from 'react';
-import {Highlighter} from '@common/highlighter';
+import Highlighter from '@common/Highlighter';
 import {WelcomeCode} from '@common/code/react/WelcomeCode';
 
 export const Welcome = () => {
