+++ src/home/pages/react/exProps/ExProps.tsx
@@ -3,7 +3,7 @@ import React from 'react';
 import {DynamicPropsFc} from './dynamicProps/DynamicPropsFc';
 import DynamicPropsClass from './dynamicProps/DynamicPropsClass';
 import RequiredProps from './RequiredProps';
-import {Highlighter} from '@common/highlighter';
+import Highlighter from '@common/Highlighter';
 import {
     RequiredPropsCode,
     DynamicPropsClassCode,
